+++ src/main/kotlin/AST/ASTNode.kt
@@ -5,4 +5,8 @@ import czechtina.GrammarToken
 open abstract class ASTNode {
 
     abstract fun toC(): String
+
+    abstract fun copy(): ASTNode
+
+    abstract fun retype(map: Map<String, String>)
 }