+++ src/main/kotlin/AST/ASTOperandNode.kt
@@ -14,8 +14,18 @@ class ASTOperandNode : ASTTypedNode {
         this.right = right
     }
 
+    override fun retype(map: Map<String, String>) {
+        left.retype(map)
+        right.retype(map)
+        expType = Compiler.calcBinaryType(left, right, operand)
+    }
+
     override fun toString(): String {
         return "'$operand', \nleft=${left.toString().replace("\n","\n\t")}, \nright=${right.toString().replace("\n","\n\t")}\n"
+    }
+
+    override fun copy(): ASTOperandNode {
+        return ASTOperandNode(operand, left.copy(), right.copy())
     }
 
     override fun toC(): String = when (operand) {