+++ src/main/kotlin/AST/ASTProgramLines.kt
@@ -12,5 +12,15 @@ class ASTProgramLines : ASTNode {
         return "Lines:\n${programLines.joinToString("").replace("\n", "\n\t")}\n"
     }
 
+    override fun copy(): ASTProgramLines {
+        return ASTProgramLines(
+            programLines.map { it.copy() }
+        )
+    }
+
+    override fun retype(map: Map<String, String>) {
+        programLines.forEach { it.retype(map) }
+    }
+
     override fun toC(): String = programLines.joinToString("\n") { it.toC() }
 }