+++ src/main/kotlin/compiler/Compiler.kt
@@ -1,6 +1,8 @@
 package compiler
 
+import AST.ASTListNode
 import AST.ASTNode
+import AST.ASTProgramNode
 import AST.ASTTypedNode
 import cz.j_jzk.klang.input.InputFactory
 import czechtina.*
@@ -12,10 +14,10 @@ object Compiler {
     val VERSION = "0.1.5"
     var compilingTo = "C"
     var definedTypes = mutableListOf<String>()
-    var definedFunctions = mutableMapOf<String, DefinedType>(
-        "printf" to DefinedType("void", false),
-        "new" to DefinedType("dynamic-void", true),
-        "predej" to DefinedType("dynamic-void", true),
+    var definedFunctions = mutableMapOf<String, DefinedFunction>(
+        "printf" to DefinedFunction("printf",DefinedType("void"), listOf(DefinedFunctionVariant("printf", listOf("string"))), virtual = true),
+        "new" to DefinedFunction("new",DefinedType("dynamic-void", true), listOf(DefinedFunctionVariant("malloc", listOf("int"))), virtual = true),
+        "predej" to DefinedFunction("predej",DefinedType("dynamic-void"), listOf(DefinedFunctionVariant("", listOf("dynamic"))), virtual = true),
     )
     var variables = mutableListOf(mutableMapOf<String, DefinedType>())
     var grammar: Map<GrammarToken,String> = C
@@ -39,6 +41,8 @@ object Compiler {
             throw Exception("Variable $varName is defined as type")
         if (definedFunctions.containsKey(varName))
             throw Exception("Variable $varName is defined as function")
+        if (variables.any { it.containsKey(varName) })
+            throw Exception("Variable $varName is already defined")
         return true
     }
 
@@ -62,18 +66,20 @@ object Compiler {
         return ""
     }
 
-    fun scopePop(write:Boolean = false): String {
-        var retVal = "\n\t"
+    fun scopePop(write:Boolean = false, init: String = "\n\t", end: String = ""): String {
+        var retVal = init
         if (variables.size == 1)
             throw Exception("Cant pop global scope")
         variables[variables.size -1].forEach {
-            if (it.value.isHeap)
+            if (it.value.isHeap && !it.value.dealocated)
                 retVal += "free(${it.key});\n\t"
         }
         variables.removeAt(variables.size - 1)
+        if (retVal != init)
+            retVal += "\n\t"
         if (!write)
             return ""
-        return retVal.substringBeforeLast("\n")
+        return retVal.substringBeforeLast("\n") + end
     }
 
 
@@ -97,6 +103,10 @@ object Compiler {
     }
 
     fun calcBinaryType(left: ASTTypedNode,  right: ASTTypedNode, operand: String): String {
+        if (left.getType().contains("*"))
+            return left.getType()
+        if (right.getType().contains("*"))
+            return right.getType()
         if (operand == "=")
             return right.getType()
 
@@ -107,7 +117,7 @@ object Compiler {
             val minW = minOf(leftWeight, rightWeight)
 
             if (maxW == 10 && minW == 2)
-                throw Exception("cant do this operation with pointer")
+                throw Exception("cant do $operand operation with pointer")
 
             if (operand == "%" && listOf(leftWeight, rightWeight).any { it == 3 || it == 4 })
                 throw Exception("Modulo cant be made with floating point number")
@@ -120,11 +130,6 @@ object Compiler {
             return right.getType()
 
         } catch (e: Exception) {
-            println("------------")
-            println(this)
-            println("operand: $operand")
-            println("left: $left ${left.getType()}")
-            println("right: $right ${right.getType()}")
             throw Exception("Error in calcBinaryType: ${e.message}")
         }
     }
@@ -133,13 +138,20 @@ object Compiler {
         return "Compiler(compilingTo='$compilingTo', definedTypes=$definedTypes, definedFunctions=$definedFunctions, variables=$variables,)"
     }
 
+    fun isAnyUsedFunctionUndefined(): Boolean {
+        for (function in definedFunctions)
+            if (function.value.variants.any { !it.defined && it.timeUsed > 0 })
+                return true
+        return false
+    }
 
     fun compileFile(path: String, args: Array<String>) {
-        val code = Preprocessor.preprocess(path)
+        var code = Preprocessor.preprocess(path)
         var withoutExtension = path.substring(0, path.length - 3)
         val czechtina = czechtinaLesana()
 
-        val tree = czechtina.parse(InputFactory.fromString(code, "code")) as ASTNode
+
+        val tree = czechtina.parse(InputFactory.fromString(code, "code")) as ASTProgramNode
 
         if (args.any() { it == "--show-tree" }) {
             println(tree.toString())
@@ -149,6 +161,56 @@ object Compiler {
         variables.add(mutableMapOf())
 
         var cCode = tree.toC()
+
+
+
+        while (isAnyUsedFunctionUndefined()) {
+            for (function in definedFunctions){
+                val fce = function.value
+                if (fce.virtual)
+                    continue
+                val variants = listOf(fce.variants).flatten().filter { !it.defined && it.timeUsed > 0 }
+                for (variant in variants){
+                    val functionASTref = tree.functions.find { it.name == function.key }
+                        ?: throw Exception("Function ${function.key} not found")
+                    Compiler.scopePush()
+                    var functionAST = functionASTref.copy()
+                    Compiler.scopePop()
+                    val paramsTypes = mutableListOf<String>()
+                    for (param in functionAST.parameters)
+                        if (param is ASTTypedNode)
+                            paramsTypes.add(param.getType())
+                    val abstractIndex = fce.validateParams(paramsTypes)
+                    var retypeMap = mutableMapOf<String,String>()
+                    for (i in 0 until variant.params.size){
+                        var old =fce.variants[abstractIndex].params[i]
+                        var new = variant.params[i]
+
+                        if (old.contains("pointer")){
+                            val newOld = old.replace("pointer", "dynamic")
+                            if (newOld == new) {
+                                retypeMap += mapOf(old to newOld)
+                                continue
+                            }
+                        }
+
+
+                        if (!old.contains("*"))
+                            continue
+                        if (old.contains("-"))
+                            old = old.split("-").find { it.contains("*") }!!
+
+
+                        retypeMap += mapOf(old to new)
+                    }
+                    functionAST.name = variant.translatedName
+                    functionAST.retype(retypeMap)
+                    cCode = cCode.replace("//${function.key}_Declaration_CZECHTINA ANCHOR", "//${function.key}_Declaration_CZECHTINA ANCHOR\n${functionAST.toCDeclarationNoSideEffect()}")
+                    cCode = cCode.replace("//${function.key}_CZECHTINA ANCHOR", "//${function.key}_CZECHTINA ANCHOR\n${functionAST.toCNoSideEffect()}")
+                    variant.defined = true
+                }
+            }
+        }
 
         if (Compiler.compilingTo == "CZ") {
             cCode = "#define \"czechtina.h\"\n$cCode"
@@ -163,6 +225,8 @@ object Compiler {
         }
 
         cCode = cCode.replace("#\$#CZECHTINAMEZERA\$#\$", " ")
+
+        cCode = cCode.lines().filter { !it.contains("CZECHTINA ANCHOR") }.joinToString("\n")
 
         cCode= "// Czechtina ${Compiler.VERSION}\n$cCode"