+++ src/main/kotlin/czechtina/lesana/czechtinaLesana.kt
@@ -2,6 +2,7 @@ package czechtina.lesana
 
 import AST.*
 import compiler.Compiler
+import cz.j_jzk.klang.lesana.LesanaBuilder
 import cz.j_jzk.klang.lesana.lesana
 import cz.j_jzk.klang.parse.NodeID
 import czechtina.*
@@ -12,6 +13,7 @@ fun czechtinaLesana() = lesana<ASTNode> {
     val variables = NodeID<ASTVariableNode>("variables")
     val operands = NodeID<String>("operands")
     val endOfLine = re(";|kafe")
+    val konec = re("konec")
     val main = NodeID<ASTFunctionNode>("main")
     val tFunction = NodeID<ASTFunctionNode>("function")
     val blockCode = NodeID<ASTUnaryNode>("blockCode")
@@ -23,7 +25,6 @@ fun czechtinaLesana() = lesana<ASTNode> {
     val import = NodeID<ASTUnaryNode>("import")
     val r_expression = include(expression(variables, types))
     val program = NodeID<ASTProgramNode>("program")
-
 
 
     typeDefinition to def(
@@ -49,117 +50,15 @@ fun czechtinaLesana() = lesana<ASTNode> {
         re(">"))
     { (_, _, t2, _) -> ASTUnaryNode(ASTUnaryTypes.TYPE_POINTER, t2, "pointer-${t2.getType()}") }
 
-    types to def(re(cAndCzechtinaRegex(Alltypes))) { ASTUnaryNode(ASTUnaryTypes.TYPE, it.v1, cTypeFromCzechtina(it.v1)) }
+    types to def(re(cAndCzechtinaRegex(Alltypes)+"|T[0-9]*")) { ASTUnaryNode(ASTUnaryTypes.TYPE, if (it.v1.contains("T")) "*${it.v1}" else it.v1, if (it.v1.contains("T")) "*${it.v1}" else cTypeFromCzechtina(it.v1) ) }
 
     operands to def(re(cAndCzechtinaRegex(listOf(GrammarToken.OPERATOR_ASSIGN)))) { it.v1 }
 
     // FOR LOOP
-
-    line to def (
-        re(cAndCzechtinaRegex(listOf(GrammarToken.KEYWORD_FOR))),
-        variables,
-        re(czechtina[GrammarToken.KEYWORD_VAR_DEFINITION]!!),
-        types,
-        re(czechtina[GrammarToken.OPERATOR_ITERATE]!!),
-        r_expression,
-        re(czechtina[GrammarToken.KEYWORD_RANGE_DEFINITION]!!),
-        r_expression,
-        blockCode
-    )
-    { (_, v, _, t, _, min, def, max, block) -> ASTForNode(v, t, min, def, max, block) }
-
-    line to def (
-        re(cAndCzechtinaRegex(listOf(GrammarToken.KEYWORD_FOR))),
-        variables,
-        re(czechtina[GrammarToken.OPERATOR_ITERATE]!!),
-        r_expression,
-        re(czechtina[GrammarToken.KEYWORD_RANGE_DEFINITION]!!),
-        r_expression,
-        blockCode
-    )
-    { (_, v, _, min, def, max, block) -> ASTForNode(v, ASTUnaryNode(ASTUnaryTypes.TYPE, min.getType(), min.getType()), min, def, max, block) }
-
-    line to def(
-        re(cAndCzechtinaRegex(listOf(GrammarToken.KEYWORD_FOR))),
-        line,
-        r_expression,
-        endOfLine,
-        r_expression,
-        blockCode
-    ) {(_, begin,cond,_,step, block) -> ASTForNode(begin, cond, step, block) }
-
-    // inline for
-
-    line to def (
-        re(cAndCzechtinaRegex(listOf(GrammarToken.KEYWORD_FOR))),
-        variables,
-        re(czechtina[GrammarToken.KEYWORD_VAR_DEFINITION]!!),
-        types,
-        re(czechtina[GrammarToken.OPERATOR_ITERATE]!!),
-        r_expression,
-        re(czechtina[GrammarToken.KEYWORD_RANGE_DEFINITION]!!),
-        r_expression,
-        re(czechtina[GrammarToken.OPERATOR_ITERATE]!!),
-        line
-    )
-    { (_, v, _, t, _, min, def, max, _,block) -> ASTForNode(v, t, min, def, max, ASTUnaryNode(ASTUnaryTypes.NEW_LINE, block)) }
-
-    line to def (
-        re(cAndCzechtinaRegex(listOf(GrammarToken.KEYWORD_FOR))),
-        variables,
-        re(czechtina[GrammarToken.OPERATOR_ITERATE]!!),
-        r_expression,
-        re(czechtina[GrammarToken.KEYWORD_RANGE_DEFINITION]!!),
-        r_expression,
-        re(czechtina[GrammarToken.OPERATOR_ITERATE]!!),
-        line
-    )
-    { (_, v, _, min, def, max, _,block) -> ASTForNode(v, ASTUnaryNode(ASTUnaryTypes.TYPE, min.getType(), min.getType()), min, def, max, ASTUnaryNode(ASTUnaryTypes.NEW_LINE, block)) }
-
-    line to def(
-        re(cAndCzechtinaRegex(listOf(GrammarToken.KEYWORD_FOR))),
-        line,
-        r_expression,
-        endOfLine,
-        r_expression,
-        re(czechtina[GrammarToken.OPERATOR_ITERATE]!!),
-        line
-    ) {(_, begin,cond,_,step,_, block) -> ASTForNode(begin, cond, step,ASTUnaryNode(ASTUnaryTypes.NEW_LINE, block)) }
+    forLoops(line, variables, types, r_expression, blockCode, endOfLine)
 
 
-    varDefinition to def(
-        variables,
-        re(czechtina[GrammarToken.KEYWORD_VAR_DEFINITION]!!),
-        re(czechtina[GrammarToken.TYPE_ARRAY]!!),
-        re("<"),
-        types,
-        re(">")
-    ) { (v, _,_,_, t,_) -> ASTStaticArrayDefinitionNode(t, v.addType("array-${t.getType()}-"), "") }
-
-    varDefinition to def(
-        variables,
-        re(czechtina[GrammarToken.KEYWORD_VAR_DEFINITION]!!),
-        re(czechtina[GrammarToken.TYPE_ARRAY]!!),
-        re("<"),
-        types,
-        re(","),
-        re("[0-9]+"),
-        re(">")
-    ) { (v, _,_,_, t,_, s, _) -> ASTStaticArrayDefinitionNode(t, v.addType("array-${t.getType()}-$s"), s) }
-
-
-    varDefinition to def(
-        variables,
-        re(czechtina[GrammarToken.KEYWORD_VAR_DEFINITION]!!),
-        types
-    ) { (v, _, t) -> ASTVarDefinitionNode(v.addType(t.getType()), t) }
-
-
-    varDefinition to def(
-        variables,
-        re(czechtina[GrammarToken.KEYWORD_VAR_DEFINITION]!!),
-        variables
-    ) { (v, _, t) -> if (Compiler.definedTypes.contains(t.toC())) ASTVarDefinitionNode( v,t) else throw Exception("Variable ${t.toC()} is not defined as an type") }
+    variableDefinition(varDefinition, variables, types)
 
     line to def(
         varDefinition,
@@ -184,50 +83,8 @@ fun czechtinaLesana() = lesana<ASTNode> {
         endOfLine
     ) { ASTUnaryNode(ASTUnaryTypes.SEMICOLON, ASTUnaryNode(ASTUnaryTypes.RETURN, ASTUnaryNode(ASTUnaryTypes.LITERAL, ""))) }
 
-    // IF
 
-    line to def(
-        re(cAndCzechtinaRegex(listOf(GrammarToken.KEYWORD_IF))),
-        r_expression,
-        blockCode
-    ) { (_, exp, block) -> ASTBinaryNode(ASTBinaryTypes.FLOW_CONTROL, ASTUnaryNode(ASTUnaryTypes.IF,exp), block)}
-
-    line to def(
-        re(cAndCzechtinaRegex(listOf(GrammarToken.KEYWORD_IF))),
-        r_expression,
-        re(czechtina[GrammarToken.OPERATOR_ITERATE]!!),
-        line
-    ) { (_, exp,_, block) -> ASTBinaryNode(ASTBinaryTypes.FLOW_CONTROL, ASTUnaryNode(ASTUnaryTypes.IF,exp), ASTUnaryNode(ASTUnaryTypes.NEW_LINE, block) )}
-
-    // ELSE IF
-
-    line to def(
-        re(cAndCzechtinaRegex(listOf(GrammarToken.KEYWORD_ELSE))),
-        re(cAndCzechtinaRegex(listOf(GrammarToken.KEYWORD_IF))),
-        r_expression,
-        blockCode
-    ) { (_,_,exp, block) -> ASTBinaryNode(ASTBinaryTypes.FLOW_CONTROL, ASTUnaryNode(ASTUnaryTypes.ELSE_IF, exp), block) }
-
-    line to def(
-        re(cAndCzechtinaRegex(listOf(GrammarToken.KEYWORD_ELSE))),
-        re(cAndCzechtinaRegex(listOf(GrammarToken.KEYWORD_IF))),
-        r_expression,
-        re(czechtina[GrammarToken.OPERATOR_ITERATE]!!),
-        line
-    ) { (_,_,exp,_, block) -> ASTBinaryNode(ASTBinaryTypes.FLOW_CONTROL, ASTUnaryNode(ASTUnaryTypes.ELSE_IF, exp), ASTUnaryNode(ASTUnaryTypes.NEW_LINE, block)) }
-
-    // ELSE
-
-    line to def(
-        re(cAndCzechtinaRegex(listOf(GrammarToken.KEYWORD_ELSE))),
-        blockCode
-    ) { (_, block) -> ASTBinaryNode(ASTBinaryTypes.FLOW_CONTROL, ASTUnaryNode(ASTUnaryTypes.ELSE, ""), block) }
-
-    line to def(
-        re(cAndCzechtinaRegex(listOf(GrammarToken.KEYWORD_ELSE))),
-        re(czechtina[GrammarToken.OPERATOR_ITERATE]!!),
-        blockCode
-    ) { (_, _, block) -> ASTBinaryNode(ASTBinaryTypes.FLOW_CONTROL, ASTUnaryNode(ASTUnaryTypes.ELSE, ""), ASTUnaryNode(ASTUnaryTypes.NEW_LINE, block)) }
+    flowControl(line, r_expression, blockCode)
 
 
 
@@ -243,144 +100,21 @@ fun czechtinaLesana() = lesana<ASTNode> {
         ASTUnaryNode(ASTUnaryTypes.CURLY, it.v2)
     }
 
-
     // MAIN FUNCTION
     main to def(
         re("main"), blockCode
     )
     {
-        ASTFunctionNode(
-            ASTUnaryNode(ASTUnaryTypes.TYPE, "cele"), it.v1,
-            emptyList(), it.v2
-        )
+       ASTFunctionNode(
+           ASTUnaryNode(ASTUnaryTypes.TYPE, "cele"), it.v1,
+           emptyList(), it.v2
+       )
     }
 
+    blockFunction(tFunction, varDefinition, types, programLines, listableDefinition)
+    inlineTypedFunction(tFunction, varDefinition, types, r_expression, konec, listableDefinition)
 
-
-    tFunction to def(
-        re("[a-zA-Z][a-zA-Z0-9]*"),
-        varDefinition,
-        re("{"),
-        types,
-        programLines,
-        re("}")
-    )
-    { (funName, varDef, _, retType, lines, _) ->
-        ASTFunctionNode(
-            retType, funName,
-            listOf(varDef), ASTUnaryNode(ASTUnaryTypes.CURLY_UNSCOPE, lines)
-        )
-    }
-    tFunction to def(
-        re("[a-zA-Z][a-zA-Z0-9]*"),
-        listableDefinition,
-        re("{"),
-        types,
-        programLines,
-        re("}")
-    )
-    { (funName, varDefs, _, retType, lines, _) ->
-        ASTFunctionNode(
-            retType, funName,
-            varDefs.nodes,  ASTUnaryNode(ASTUnaryTypes.CURLY_UNSCOPE, lines)
-        )
-    }
-
-    // inline function
-    tFunction to def(
-        re("[a-zA-Z][a-zA-Z0-9]*"),
-        varDefinition,
-        re(czechtina[GrammarToken.OPERATOR_ITERATE]!!),
-        types,
-        r_expression,
-    )
-    { (funName, varDef, _, retType, line) ->
-        ASTFunctionNode(
-            retType,
-            funName,
-            listOf(varDef),
-            ASTUnaryNode(ASTUnaryTypes.CURLY_UNSCOPE,  ASTUnaryNode(ASTUnaryTypes.SEMICOLON, ASTUnaryNode(ASTUnaryTypes.RETURN, line)))
-        )
-    }
-
-    // without type
-    tFunction to def(
-        re("[a-zA-Z][a-zA-Z0-9]*"),
-        varDefinition,
-        re(czechtina[GrammarToken.OPERATOR_ITERATE]!!),
-        r_expression,
-    )
-    { (funName, varDef, _, line) ->
-        ASTFunctionNode(
-            ASTUnaryNode(ASTUnaryTypes.TYPE, line.getType()),
-            funName,
-            listOf(varDef),
-            ASTUnaryNode(ASTUnaryTypes.CURLY_UNSCOPE,  ASTUnaryNode(ASTUnaryTypes.SEMICOLON, ASTUnaryNode(ASTUnaryTypes.RETURN, line)))
-        )
-    }
-
-    // multiple params
-    tFunction to def(
-        re("[a-zA-Z][a-zA-Z0-9]*"),
-        listableDefinition,
-        re(cAndCzechtinaRegex(listOf(GrammarToken.OPERATOR_ASSIGN))),
-        types,
-        r_expression,
-    )
-    { (funName, varDefs, _, retType, line) ->
-        ASTFunctionNode(
-            retType,
-            funName,
-            varDefs.nodes,
-            ASTUnaryNode(ASTUnaryTypes.CURLY_UNSCOPE, ASTUnaryNode(ASTUnaryTypes.SEMICOLON, ASTUnaryNode(ASTUnaryTypes.RETURN, line)))
-        )
-    }
-
-    tFunction to def(
-        re("[a-zA-Z][a-zA-Z0-9]*"),
-        listableDefinition,
-        re(cAndCzechtinaRegex(listOf(GrammarToken.OPERATOR_ASSIGN))),
-        r_expression,
-    )
-    { (funName, varDefs, _, line) ->
-        ASTFunctionNode(
-            ASTUnaryNode(ASTUnaryTypes.TYPE, line.getType()),
-            funName,
-            varDefs.nodes,
-            ASTUnaryNode(ASTUnaryTypes.CURLY_UNSCOPE, ASTUnaryNode(ASTUnaryTypes.SEMICOLON, ASTUnaryNode(ASTUnaryTypes.RETURN, line)))
-        )
-    }
-
-    // no params
-    tFunction to def(
-        re("[a-zA-Z][a-zA-Z0-9]*"),
-        re(czechtina[GrammarToken.OPERATOR_ITERATE]!!),
-        types,
-        r_expression,
-    )
-    { (funName, _, retType, line) ->
-        ASTFunctionNode(
-            retType,
-            funName,
-            listOf(),
-            ASTUnaryNode(ASTUnaryTypes.CURLY_UNSCOPE,  ASTUnaryNode(ASTUnaryTypes.SEMICOLON, ASTUnaryNode(ASTUnaryTypes.RETURN, line)))
-        )
-    }
-
-    // without type
-    tFunction to def(
-        re("[a-zA-Z][a-zA-Z0-9]*"),
-        re(czechtina[GrammarToken.OPERATOR_ITERATE]!!),
-        r_expression,
-    )
-    { (funName, _, line) ->
-        ASTFunctionNode(
-            ASTUnaryNode(ASTUnaryTypes.TYPE, line.getType()),
-            funName,
-            listOf(),
-            ASTUnaryNode(ASTUnaryTypes.CURLY_UNSCOPE,  ASTUnaryNode(ASTUnaryTypes.SEMICOLON, ASTUnaryNode(ASTUnaryTypes.RETURN, line)))
-        )
-    }
+    inlineFunction(tFunction, varDefinition, r_expression, konec, listableDefinition)
 
 
 
@@ -388,7 +122,6 @@ fun czechtinaLesana() = lesana<ASTNode> {
     programLines to def(line, programLines) { ASTProgramLines(listOf(it.v1) + it.v2.programLines) }
     programLines to def(line) { ASTProgramLines(listOf(it.v1)) }
 
-    //import to def(re(czechtinaRegex(listOf(GrammarToken.KEYWORD_IMPORT))), re("^(.+)\\\\/([^\\\\/]+)\$")) { ASTUnaryNode(ASTUnaryTypes.IMPORT, it.v2) }
     import to def(re(czechtinaRegex(listOf(GrammarToken.KEYWORD_IMPORT_C))), re("[a-zA-Z][a-zA-Z0-9]*")) {
         ASTUnaryNode(
             ASTUnaryTypes.IMPORT_C,
@@ -403,7 +136,6 @@ fun czechtinaLesana() = lesana<ASTNode> {
     program to def(program, typeDefinition) { (program, typ) -> program.appendTypeDefinition(typ) }
     program to def(tFunction, program) { (func, program) -> program.appendFunction(func) }
     program to def(program, tFunction) { (program, func) -> program.appendFunction(func) }
-
     program to def(main) { ASTProgramNode(listOf(), listOf(), it.v1) }
 
     line to def (blockCode) { it.v1 }